requirements.txt
Flask
pyjwt
dotenv
requests
